requirements.txt
discord.py>=2.6
aiohttp
motor
pymongo
flask
requests
python-dotenv